+++ changes.md
@@ -1,6 +1,7 @@
 # Zealot
 
  - Using the native curl cookie jar
+ - Added `HttpSession.options` 
 
 ## Release v0.8.2
 LTS: 10/04/2026